Split rendered tooling contract tests
Move the large tooling contract helper and rendered template test module into
focused packages while preserving the existing public helper imports. Tighten
the generated Makefile doctest condition to the library flavour and use a
direct `tests.utilities` import in helper tests.
