+++ tests/test_template/__init__.py
@@ -1,0 +1,1 @@
+"""Rendered template test package."""
